refactor: Move SharingBanner into components
