+++ packages/cozy-sharing/src/components/SharingBanner/components/CozyHomeLinkIcon.jsx
