+++ packages/cozy-sharing/src/components/SharingBanner/components/PublicBanner.spec.jsx
