+++ packages/cozy-sharing/src/components/SharingBanner/components/SharingBanner.jsx
