+++ packages/cozy-sharing/src/components/SharingBanner/helpers/QueryParameter.js
