+++ packages/cozy-sharing/src/components/SharingBanner/helpers/QueryParameter.spec.js
