+++ packages/cozy-sharing/src/components/SharingBanner/hooks/useSharingInfos.spec.js
