+++ packages/cozy-sharing/src/components/SharingBanner/index.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 
 import { SharingBanner } from './components/SharingBanner'
 import { useSharingInfos } from './hooks/useSharingInfos'
-import withLocales from '../withLocales'
+import withLocales from '../../withLocales'
 
 const Plugin = ({ previewPath }) => {
   const sharingInfos = useSharingInfos(previewPath)
