+++ packages/cozy-sharing/src/components/SharingBanner/test/AcceptingSharingContext.jsx
