+++ packages/cozy-sharing/src/components/SharingBanner/test/AppLike.jsx
@@ -11,8 +11,8 @@ import FabProvider from './FabProvider'
 import { ModalContext } from './ModalContext'
 import { RouterContext } from './RouterContext'
 import { ThumbnailSizeContextProvider } from './ThumbnailSizeContext'
-import langEn from '../../../locales/en.json'
-import SharingContext from '../../context'
+import langEn from '../../../../locales/en.json'
+import SharingContext from '../../../context'
 
 const mockStore = createStore(() => ({
   mobile: {
