+++ packages/cozy-sharing/src/components/SharingBanner/test/FabProvider.jsx
