+++ packages/cozy-sharing/src/components/SharingBanner/test/ModalContext.jsx
