+++ packages/cozy-sharing/src/components/SharingBanner/test/ThumbnailSizeContext.jsx
