+++ packages/cozy-sharing/src/index.jsx
@@ -19,8 +19,8 @@ export { ShareButton } from './ShareButton'
 export { ShareModal } from './components/ShareModal/ShareModal'
 export { RefreshableSharings } from './RefreshableSharings'
 export { CozyPassFingerprintDialogContent } from './components/CozyPassFingerprintDialogContent'
-export { SharingBannerPlugin } from './SharingBanner'
-export { useSharingInfos } from './SharingBanner/hooks/useSharingInfos'
+export { SharingBannerPlugin } from './components/SharingBanner'
+export { useSharingInfos } from './components/SharingBanner/hooks/useSharingInfos'
 export { ConfirmTrustedRecipientsDialog } from './ConfirmTrustedRecipientsDialog'
 export { ShareButtonWithRecipients } from './ShareButtonWithRecipients'
 export { DOCUMENT_TYPE } from './helpers/documentType'
